move alloc to own module
